Extract user managemnt code into a separate helper module
